64-bit optimization: Convert Platform memory functions to size_t
Convert file I/O functions from unsigned long to size_t for proper
64-bit support and enable handling of files larger than 4GB.

Platform.h interface changes:
- GetFileSize(): unsigned long → size_t
- CopyFileToBuffer(): unsigned long → size_t parameter

Windows implementation improvements (PlatformWin32.cpp):
- Replace fseek/ftell with _fseeki64/_ftelli64 for 64-bit file I/O
- Properly handle files larger than 4GB (previous limit: 2^32 bytes)

Unix implementation (PlatformUnix.cpp):
- Updated signatures for consistency

Call site updates:
- PreprocessParser.cpp: size_t for shader file loading
- GpuShaderManager.cpp: size_t for shader file operations
- CacheFileDateFix.cpp: size_t for cache file validation
- LuxDebugHandler.cpp: size_t for batch load file operations
- WidgetBaseClasses.h/cpp: GetHumanReadableSize() parameter type

Bug fix in WidgetBaseClasses.cpp:
- Added explicit cast to unsigned long for ToStringW() to resolve
  function overload ambiguity with size_t parameter

Updated README.md to document 64-bit optimizations.

+++ HPL2/core/include/gui/WidgetBaseClasses.h
@@ -219,7 +219,7 @@ namespace hpl {
 
 		void GetFilesAndFoldersInCurrentPath(int alCategoryIndex, tWStringList& alstDestination);
 		eFileBrowserFileType GetFileTypeByName(const tWString& asFilename);
-		tWString GetHumanReadableSize(unsigned long alSize);
+		tWString GetHumanReadableSize(size_t alSize);
 
 
 		tWString InterpretPath(const tWString& asPath);
